Criado o DTO de Declaração e itens da declaração

+++ src/main/java/br/com/senai/reciclaville/models/entities/Materials.java
@@ -22,5 +22,5 @@ public class Materials {
     private String nome;
 
     @Column(nullable = false)
-    private BigDecimal percentualCompensacao;
+    private Double percentualCompensacao;
 }
